actions: Split out zerver.actions.typing.
